测试了一下，读写文件是没有问题的，但是涉及json解析就不行了

+++ ectm_app/main/inc/xm_file.h
@@ -5,8 +5,8 @@
 #include "cJSON.h"
 
 void initSpiffs(void);
-uint32_t readFile(const char *path, char *buff, uint32_t length);
-uint32_t writeFile(const char *path, char *buff, uint32_t length);
+size_t readFile(const char *path, char *buff, uint32_t length);
+size_t writeFile(const char *path, char *buff, uint32_t length);
 cJSON* getJsonNode(cJSON *root, char *keys);
 
 #endif // !_XM_FILE_H_
